pipeline/requirements.txt
pytest
pytest-cov
requests
pandas
python-dotenv
pymssql
aiohttp
certifi
asyncio
pytest-asyncio
pylint